refactor: move extractDishes function to helpers and update import path in tests
