+++ backend/src/repositories/helpers/extract-mensa-dishes.ts
